Bugfixing and code cleanup:
1. Fixed crasher when dependent attributes were encoded.
2. Improved documentation to PointCloud::AddAttribute() methods
3. Fixed some of the inconsistencies with int64_t vs size_t usage
4. Fixed potential IO problems when the input stream does not start from 0.

+++ mesh/mesh.h
@@ -56,7 +56,7 @@ class Mesh : public PointCloud {
 
   // Sets the total number of faces. Creates new empty faces or deletes
   // existings ones if necessary.
-  void SetNumFaces(int64_t num_faces) { faces_.resize(num_faces, Face()); }
+  void SetNumFaces(size_t num_faces) { faces_.resize(num_faces, Face()); }
 
   FaceIndex::ValueType num_faces() const { return faces_.size(); }
   const Face &face(FaceIndex face_id) const {
